Fix Boxer closing the main application windwo on macOS
When boxer::show(...) is called in the applicationDidFinishLaunching:
delegate method, the message window pops up only for a short moment and
the message window and the application main window is gone. This also
happens if boxer::show(...) is called from inside windowWillClose:.
This commit removes the two lines forcing the message windwo to stay on
top of all other windows. Somehow this fixes the issue.