Add a 'Quit' Button
If Boxer is used to show the user a fatal error dialog it is more
intuitive to show an 'Quit' button instead of an 'OK' button, since the
only option is to quit the application. This commits adds
boxer::Buttons::Quit. Since there is no such option on Windows, it shows
the 'OK' button on Windows instead.